requirements.txt
openai
rich
pyyaml
python-dotenv
asyncio
google-generativeai
tiktoken
click
